+++ maps/map.lua
@@ -59,7 +59,9 @@ return {
       offsety = 0,
       parallaxx = 1,
       parallaxy = 1,
-      properties = {},
+      properties = {
+        ["collidable"] = true
+      },
       encoding = "lua",
       chunks = {
         {